pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xchat-bot-starter-pro"
version = "0.1.0"
description = "Production-grade starter kit for building XChat bots on the X Activity API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "xchat-bot-starter-pro contributors" }]
keywords = ["xchat", "twitter", "x", "bot", "dm", "activity-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "typer>=0.12",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "tenacity>=8.3",
    "cryptography>=42.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "anyio>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx",
    "ruff>=0.4",
    "pyright>=1.1",
    "pip-audit>=2.7",
]

[project.scripts]
xchat = "xchat_bot.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/xchat_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: pure unit tests, no I/O",
    "integration: may use httpx or event loop",
    "slow: longer running tests",
]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "ANN"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["src"]
exclude = ["tests"]

[tool.coverage.run]
source = ["src/xchat_bot"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
